# docs/swagger.yaml
basePath: /
definitions:
...
  ds.Users:
    properties:
      id:
        type: integer
      login:
        type: string
      mod_status:
        $ref: '#/definitions/role.Role'
      password:
        type: string
    type: object
  handler.loginReq:
    properties:
      guest:
        type: boolean
      login:
        type: string
      password:
        type: string
    required:
    - guest
    - login
    - password
    type: object
  handler.loginResp:
    properties:
      access_token:
        type: string
      expires_in:
        type: string
      login:
        type: string
      role:
        $ref: '#/definitions/role.Role'
      token_type:
        type: string
    required:
    - access_token
    - expires_in
    - login
    - role
    - token_type
    type: object
  handler.registerReq:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  handler.registerResp:
    properties:
      ok:
        type: boolean
    required:
    - ok
    type: object
  role.Role:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      Guest: "0"
      Moderator: "2"
      User: "1"
    x-enum-descriptions:
    - "0"
    - "1"
    - "2"
    x-enum-varnames:
    - Guest
    - User
    - Moderator
host: 127.0.0.1
info:
  contact:
    email: Barinovartem383@gmail.com
    name: API Support
    url: https://vk.com/aaaaaaeaaaa
  description: Bmstu Open IT Platform
  license:
    name: AS IS (NO WARRANTY)
  title: Energy Calculation System
  version: "1.0"
paths:
...
  /user/edit:
    put:
      description: Update current user profile (login and password)
      parameters:
      - description: User data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Users'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: very very friendly response
      parameters:
      - description: user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.loginResp'
        "403":
          description: Forbidden
      summary: Login the specified user
      tags:
      - User
  /user/logout:
    post:
      description: Logout user and blacklist JWT token
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User logout
      tags:
      - User
  /user/profile:
    get:
      description: Get current user profile information
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user profile
      tags:
      - User
  /user/register:
    post:
      description: Register new user account
      parameters:
      - description: Registration data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.registerReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.registerResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User registration
      tags:
      - User
schemes:
- http
swagger: "2.0"